% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_data.R
\name{aggregate_itt_data}
\alias{aggregate_itt_data}
\title{aggregate_itt_data}
\usage{
aggregate_itt_data(filepath)
}
\arguments{
\item{filepath}{Path to a specific data-set}
}
\value{
Aggregated data-set
}
\description{
function that aggregates data-sets and adds IIT passed col
}
