% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_data.R
\name{aggregate_response_ratings}
\alias{aggregate_response_ratings}
\title{aggregate_response_ratings}
\usage{
aggregate_response_ratings(data_in)
}
\arguments{
\item{data_in}{Data frame to be aggregated}
}
\value{
Aggregated data frame
}
\description{
Summaries data-set by calculating the mean response rating for
each argument
}
