% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_data.R
\name{extract_condition}
\alias{extract_condition}
\title{extract_condition}
\usage{
extract_condition(filepath)
}
\arguments{
\item{filepath}{Path to a specific data-set}
}
\value{
Condition for the data-set as a factor
}
\description{
Gets the condition for the data-set from the file path
}
