% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_data.R
\name{get_colspec}
\alias{get_colspec}
\title{get_colspec}
\usage{
get_colspec()
}
\value{
Column specification for columns to be loaded from data
}
\description{
function detailing the columns to be loaded form the data-sets
and their data types
}
