% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_data.R
\name{get_data}
\alias{get_data}
\title{get_data}
\usage{
get_data(label, path = "data/")
}
\arguments{
\item{label}{String indicating which data-set(s) to get}
}
\value{
Returns data-frame ready to be presented in Shiny app
}
\description{
Loads and prepares data for a specific label
}
