% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_neaten_headings.R
\name{neaten_headings}
\alias{neaten_headings}
\title{neaten_headings}
\usage{
neaten_headings(column_names)
}
\arguments{
\item{column_names}{A string or vector of strings}
}
\value{
A string or vector of strings
}
\description{
`neaten_headings` is to neaten the headings of a data frame in preparation
for displaying on a Shiny App
}
\details{
Strings are neatened by removing any leading/following underscores, replacing
any sequence of underscores within the string with a single space, putting
the rest of the sting to title case, and ensuring that ITT remains
capitalised
}
