% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_data.R
\name{wrangle_raw_itt_data}
\alias{wrangle_raw_itt_data}
\title{wrangle_raw_ITT_data}
\usage{
wrangle_raw_itt_data(filepath)
}
\arguments{
\item{filepath}{Path to a specific data-set}
}
\value{
Cleaned and aggregated data-sets
}
\description{
function that loads, cleans and aggregates individual data-sets
before they are joined
}
