% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_data.R
\name{aggregate_itt_data}
\alias{aggregate_itt_data}
\title{aggregate_itt_data}
\usage{
aggregate_itt_data(raw_data)
}
\arguments{
\item{raw_data}{unaggregated data-set}
}
\value{
Aggregated data-set
}
\description{
function that aggregates data-sets and adds IIT passed col
}
